+++ CHANGELOG.md
@@ -14,8 +14,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds GLOBE AirFRANS example case (`examples/cfd/external_aerodynamics/globe/airfrans`)
 - Adds automatic support for `FSDP` and/or `ShardTensor` models in checkpoint save/load
   functionality
+- PhysicsNeMo-Mesh now supports conversion from PyVista/VTK/VTU meshes that may
+  contain polyhedral cells.
+- In PhysicsNeMo-Mesh, adds `Mesh.to_point_cloud()`, `.to_edge_graph()`, and
+  `.to_dual_graph()` methods. These allow Mesh conversion to 0D point clouds, 1D
+  edge graphs, and 1D dual graphs, respectively, when connectivity information
+  is not needed.
 
 ### Changed
+
+- Improved crash recipe with configurable stats directory.
 
 ### Deprecated